code review : 투표 제목도 하나의 atom으로 관리하는 방법으로 변경
