+++ src/components/MakerVoting/index.ts
@@ -1,3 +1,3 @@
+export { default as MakerVotingLayout } from '../Layout/MakerVotingLayout';
 export { default as ImageInput } from './ImageInput';
-export { default as MakerVotingLayout } from './MakerVotingLayout';
 export { default as TitleInput } from './TitleInput';
